Show the composition of combinations.

A function to show the composition of combinations of interest.

Usage

show_markers(markers_table, selected_combinations)

Arguments

markers_table

a data.frame with combinations returned by combi().

selected_combinations

a numeric vector that specifies the combinations of interest.

Value

a data.frame containing the selected combinations and their composing markers.

Examples

demo_data # combiroc built-in demo data (proteomics data from Zingaretti et al. 2012 - PMC3518104)
#>     Patient.ID Class Marker1 Marker2 Marker3 Marker4 Marker5
#> 1         AIH1     A     438     187     197     298     139
#> 2         AIH2     A     345     293     134     523     335
#> 3         AIH3     A     903     392     300    1253       0
#> 4         AIH4     A     552     267     296     666      22
#> 5         AIH5     A    1451     760     498     884     684
#> 6         AIH6     A     497     260     175     640     572
#> 7         AIH7     A    3878    2858     510    1518     348
#> 8         AIH8     A     740     368     266    1391     647
#> 9         AIH9     A     538     412     282    1071     474
#> 10       AIH10     A    1194    1065     442    1530     816
#> 11       AIH11     A     988     817     394    1626    1028
#> 12       AIH12     A     261     153     528     449     237
#> 13       AIH13     A     312     181      58     408     661
#> 14       AIH14     A     499     302     181    1088     568
#> 15       AIH15     A     998     431      75     397     519
#> 16       AIH16     A    6485    3704    8834    2595    1570
#> 17       AIH17     A    8584    2662   13178    3390    2757
#> 18       AIH18     A    1696    1029     369     468    2422
#> 19       AIH19     A     774     505     251    1014     517
#> 20       AIH20     A     433     330      92    1266     743
#> 21       AIH21     A    1207    1501      57    1847     741
#> 22       AIH22     A     225     849     125     330     168
#> 23       AIH23     A    1327     485     366     563     878
#> 24       AIH24     A      76     582     601     115      73
#> 25       AIH25     A     586     256      53     710      74
#> 26       AIH26     A     263     196     124     798      95
#> 27       AIH27     A     146     568      64     125      92
#> 28       AIH28     A     537     620      82     350      53
#> 29       AIH29     A      52      41     588     346      50
#> 30       AIH30     A     882     138      47     601      85
#> 31       AIH31     A     462     304      81     378     208
#> 32       AIH32     A     372     149       0     460     156
#> 33       AIH33     A    1964     875     404    1883    1021
#> 34       AIH34     A    2809    1633     451    1042     608
#> 35       AIH35     A     994     505     303    1699     783
#> 36       AIH36     A     361     185     115     466     520
#> 37       AIH37     A     275     384      45     329     466
#> 38       AIH38     A     947     622      16     236     728
#> 39       AIH39     A     405     133      32     284     358
#> 40       AIH40     A     922     689      68     105     643
#> 41     no AIH1     B     510      28       0     332       0
#> 42     no AIH2     B     205       0       0     332       0
#> 43     no AIH3     B      44       0       0     332      21
#> 44     no AIH4     B      49      38       7     178      48
#> 45     no AIH5     B      26      33       6     131      32
#> 46     no AIH6     B      24      26       8     206      28
#> 47     no AIH7     B     243     373     109     631     477
#> 48     no AIH8     B      99     111      51     193     197
#> 49     no AIH9     B      85      55      71     811     321
#> 50    no AIH10     B     146     108      46     173     171
#> 51    no AIH11     B       0     113       0       0      56
#> 52    no AIH12     B     151       0       0     360       0
#> 53    no AIH13     B       0       0       0      15      21
#> 54    no AIH14     B       0       5       0     147      53
#> 55    no AIH15     B     118      69      13     748      97
#> 56    no AIH16     B      31      29       7     223      34
#> 57    no AIH17     B      72      53      12     376      85
#> 58    no AIH18     B     137     151      20     662     131
#> 59    no AIH19     B      78      58       6     368      49
#> 60    no AIH20     B     193      14      90     703     177
#> 61    no AIH21     B       0       0       0       0       0
#> 62    no AIH22     B       0      20       0      66      31
#> 63    no AIH23     B     351      70       0     478       7
#> 64    no AIH24     B     171     127      74    1202     224
#> 65    no AIH25     B      77      75      26     316      61
#> 66    no AIH26     B      51      45       2     286      44
#> 67    no AIH27     B      64      55      17     333      46
#> 68    no AIH28     B     150     185      42     975     252
#> 69    no AIH29     B     149     140      52    1087     209
#> 70    no AIH30     B     147     103      19     794      98
#> 71    no AIH31     B      86      89      44     427     144
#> 72    no AIH32     B      94      57      18     581      88
#> 73    no AIH33     B      56      41      23     244      49
#> 74    no AIH34     B      10      15       2     121      26
#> 75    no AIH35     B      46      22       0      69      24
#> 76    no AIH36     B      84      65       4     278      77
#> 77    no AIH37     B      42      44      19     306      86
#> 78    no AIH38     B      46      51      24     342      74
#> 79    no AIH39     B      73      39       9     358      41
#> 80    no AIH40     B      83     131      27    1012     124
#> 81    no AIH41     B       0     193      50     382     135
#> 82    no AIH42     B      41      94      64     464     141
#> 83    no AIH43     B       0      77      46     398      85
#> 84    no AIH44     B      98     152      91     469     301
#> 85    no AIH45     B      78      85      10     407       0
#> 86    no AIH46     B      21      16       0     536      16
#> 87    no AIH47     B      44      34      74     266      21
#> 88    no AIH48     B     144      76      63     252       0
#> 89    no AIH49     B     162     249     101     134       0
#> 90    no AIH50     B     100     169      33    1430       0
#> 91    no AIH51     B      52     159      57     681     417
#> 92    no AIH52     B      97      86      64     591      33
#> 93    no AIH53     B      73      80      20     157      51
#> 94    no AIH54     B     162     117      25     749       0
#> 95    no AIH55     B     100     220      37     729      41
#> 96    no AIH56     B     227     174      26     290       0
#> 97    no AIH57     B     136     165      23     315      40
#> 98    no AIH58     B      96     213      47     682     780
#> 99    no AIH59     B      69      59      21     629     245
#> 100   no AIH60     B      74      70      67     138     215
#> 101   no AIH61     B     129     118      44     117     769
#> 102   no AIH62     B     265     173     115    1157    1310
#> 103   no AIH63     B     308     265     290    2464     840
#> 104   no AIH64     B     276     179     203    1848     563
#> 105   no AIH65     B     173     119     194    2245     726
#> 106   no AIH66     B     354     156       0     227     109
#> 107   no AIH67     B      73      25       0     267      68
#> 108   no AIH68     B     111      80      42     384     180
#> 109   no AIH69     B     148      90      54     479     235
#> 110   no AIH70     B     229     367      12       3     595
#> 111   no AIH71     B     388     316       0       0       0
#> 112   no AIH72     B       0       5       0     300     323
#> 113   no AIH73     B       0      66      22      92       0
#> 114   no AIH74     B       0      47       0     337       0
#> 115   no AIH75     B       0     256      42     760       0
#> 116   no AIH76     B     107      61      26     391       0
#> 117   no AIH77     B     174     109      57     428     223
#> 118   no AIH78     B      42      46      13     235     319
#> 119   no AIH79     B     180     146      34     362     249
#> 120   no AIH80     B      79      66      57     378     307
#> 121   no AIH81     B     419      86      51     376     533
#> 122   no AIH82     B     448     133      95     631     122
#> 123   no AIH83     B       0     119      49     399     154
#> 124   no AIH84     B     148      90      67     285     167
#> 125   no AIH85     B      11     155      19     231       0
#> 126   no AIH86     B     165     162     167     912     405
#> 127   no AIH87     B     320     150      88     107    4114
#> 128   no AIH88     B     146      87     111     395     238
#> 129   no AIH89     B      90     102      62     112     267
#> 130   no AIH90     B     257     188     171     644     347
#> 131   no AIH91     B     106      60      63     336       0
#> 132   no AIH92     B    1223     584     640     256       0
#> 133   no AIH93     B     519     287     283       0       0
#> 134   no AIH94     B     327     240     101     102       0
#> 135   no AIH95     B     559    2142      59     279     413
#> 136   no AIH96     B     202     201      29     248       0
#> 137   no AIH97     B     212      88      45     472       0
#> 138   no AIH98     B     426     125      68      99     566
#> 139   no AIH99     B     260     135     105     285      34
#> 140  no AIH100     B     371     257      96     528      75
#> 141  no AIH101     B     254      99     105     332      13
#> 142  no AIH102     B     107      86     121     786     168
#> 143  no AIH103     B     265     209      40     328      10
#> 144  no AIH104     B     201     189      31     559       0
#> 145  no AIH105     B      79      63      38     159       2
#> 146  no AIH106     B     111      34      32     343      13
#> 147  no AIH107     B     130     185       0     577     923
#> 148  no AIH108     B      58      29      11     545       0
#> 149  no AIH109     B      49      56      34     390       0
#> 150  no AIH110     B     158     165      68     227       0
#> 151  no AIH111     B     159     205      25     151      32
#> 152  no AIH112     B     144     155      25     600       0
#> 153  no AIH113     B     111     155      83     179      65
#> 154  no AIH114     B     224     115      27     276      67
#> 155  no AIH115     B      85      85      13     123      48
#> 156  no AIH116     B     362     296      11     176      72
#> 157  no AIH117     B     273     264     122    1418     476
#> 158  no AIH118     B     425     402      86    1702     581
#> 159  no AIH119     B     315     418       0     595     305
#> 160  no AIH120     B     463     379      71     626     313
#> 161  no AIH121     B      94     413      46     400     229
#> 162  no AIH122     B     163     115     285     590     213
#> 163  no AIH123     B       0       0       0    1053     362
#> 164  no AIH124     B       0       0       0     430      26
#> 165  no AIH125     B      83      61     299     476     219
#> 166  no AIH126     B     381     303       4     266     376
#> 167  no AIH127     B     253     106     134     626     189
#> 168  no AIH128     B     571      96      31     487     423
#> 169  no AIH129     B     288     298       0     347     411
#> 170  no AIH130     B     135     149     229     864     600

combs <- combi(data= demo_data, signalthr=450, combithr=1, case_class='A')  # compute combinations

#  To show the composition of combinations of interest.

show_markers(markers_table = combs, selected_combinations = c(1,11))
#>      Combination       Composing_markers
#> 1  Combination 1         Marker1-Marker2
#> 2 Combination 11 Marker1-Marker2-Marker3