Load unclassified data.

A function to load datasets not yet classified. It's analogue to load_data() since it loads the same data type and performs the same format checks, with the exception of "Class" column that in unclassified data is missing.

Usage

load_unclassified_data(data, sep = ";", na.strings = "")

Arguments

data

the name of the file which the data are to be read from.

sep

the field separator character.

na.strings

a character vector of strings which are to be interpreted as NA values.

Value

a data frame (data.frame) containing a representation of the data in the file.

Details

The unclassified dataset to be loaded should be in text format, which can be comma, tab or semicolon separated:

The 1st column must contain unique patient/sample IDs.

From the 2nd column on, the dataset must contain numerical values that represent the signal corresponding to the markers abundance in each sample (marker-related columns).

Marker-related columns must be called with the same name of the dataset previously loaded with load_data(). Only if all the checks are passed, it reorders alphabetically the marker-related columns depending on marker names (necessary for a proper computation of combinations), and it forces "Class" as 2nd column name.

Examples

demo_unclassified_data # combiroc built-in unclassified demo data
#>    Patient.ID Marker1 Marker2 Marker3 Marker4 Marker5
#> 1       AIH33    1964     875     404    1883    1021
#> 2   no AIH126     381     303       4     266     376
#> 3       AIH12     261     153     528     449     237
#> 4   no AIH112     144     155      25     600       0
#> 5    no AIH41       0     193      50     382     135
#> 6    no AIH38      46      51      24     342      74
#> 7        AIH4     552     267     296     666      22
#> 8    no AIH32      94      57      18     581      88
#> 9       AIH20     433     330      92    1266     743
#> 10   no AIH13       0       0       0      15      21
#> 11   no AIH11       0     113       0       0      56
#> 12  no AIH114     224     115      27     276      67
#> 13  no AIH121      94     413      46     400     229
#> 14      AIH17    8584    2662   13178    3390    2757
#> 15      AIH14     499     302     181    1088     568
#> 16  no AIH106     111      34      32     343      13
#> 17   no AIH67      73      25       0     267      68
#> 18       AIH9     538     412     282    1071     474
#> 19   no AIH74       0      47       0     337       0
#> 20      AIH16    6485    3704    8834    2595    1570

# save a data.frame as a csv to be load by combiroc package
file= tempfile()
write.csv2(demo_unclassified_data, file = file, row.names = FALSE)

# To load an unclassified dataset.

demo_unclassified_data <- load_unclassified_data(data= file ,
sep = ";", na.strings="" )